+++ app/Filament/Widgets/RecentActivityWidget.php
@@ -24,6 +24,8 @@ class RecentActivityWidget extends Widget
 {
     protected static bool $isLazy = false;
 
+    protected int|string|array $columnSpan = 'full';
+
     protected string $view = 'filament.widgets.recent-activity-widget';
 
     /**